+++ reservation/src/main/java/kr/or/connect/reservation/service/DisplayService.java
@@ -4,7 +4,7 @@ import kr.or.connect.reservation.dto.DisplayInfo;
 import kr.or.connect.reservation.dto.DisplayInfoImage;
 
 public interface DisplayService {
-	DisplayInfo getDisplayInfoUsingDisplayInfoId(Integer displayInfoId);
+	DisplayInfo getDisplayInfoUsingDisplayInfoId(Long displayInfoId);
 
-	DisplayInfoImage getDisplayInfoImageUsingDisplayInfoId(Integer displayInfoId);
+	DisplayInfoImage getDisplayInfoImageUsingDisplayInfoId(Long displayInfoId);
 }
